chore: redid the file and menu structure (#331)
* chore: redid the file and menu structure

* fix: wrong URLs
