+++ flint.ui/src/components/Stepper/StepperStatic.vue
@@ -29,10 +29,10 @@
             :xl="{ span: 20 }"
             class="bg-gray-300 border border-gray-200 card-content rounded-lg"
           >
-            <SettingFilled :style="{ fontSize: '24px', marginLeft: '5px' }" />
+            <FileFilled :style="{ fontSize: '24px', marginLeft: '5px' }" />
             <div class="bg-gray-200 rounded-r-lg body-step">
-              <h2 class="font-bold text-sm">Configure</h2>
-              <p class="text-xs text-gray-600">Change configuration parameters</p>
+              <h2 class="font-bold text-sm">Upload file</h2>
+              <p class="text-xs text-gray-600">Dataset file for GCBM</p>
             </div>
           </a-col>
           <a-col class="next-icons" :xs="{ span: 24 }" :md="{ span: 4 }" :xl="{ span: 4 }">
@@ -49,10 +49,10 @@
             :xl="{ span: 20 }"
             class="bg-gray-300 border border-gray-200 card-content rounded-lg"
           >
-            <FileFilled :style="{ fontSize: '24px', marginLeft: '5px' }" />
+            <SettingFilled :style="{ fontSize: '24px', marginLeft: '5px' }" />
             <div class="bg-gray-200 rounded-r-lg body-step">
-              <h2 class="font-bold text-sm">Upload file</h2>
-              <p class="text-xs text-gray-600">Dataset file for GCBM</p>
+              <h2 class="font-bold text-sm">Configure</h2>
+              <p class="text-xs text-gray-600">Change configuration parameters</p>
             </div>
           </a-col>
           <a-col class="next-icons" :xs="{ span: 24 }" :md="{ span: 4 }" :xl="{ span: 4 }">
