+++ flint.ui/src/components/gcbm/Uploads/Classifiers.vue
@@ -1,0 +1,14 @@
+<template>
+  <div class="py-4 px-8 text-earth">
+    <a-typography-title :level="2"><span class="font-normal text-earth"> Classifiers </span></a-typography-title>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'Classifiers',
+  setup() {
+    return {}
+  }
+}
+</script>
